pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "regat"
version = "1.0.0"
description = "Reconnaissance Automation Tool for authorized web application security assessments"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "dnspython>=2.6.1"
]
authors = [
    { name = "Shiva Guru" }
]

[project.scripts]
regat = "regat.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["regat*"]